+++ package.json
@@ -94,8 +94,6 @@
     "passport-local": "^1.0.0",
     "pg": "^8.11.3",
     "prom-client": "^15.1.3",
-    "puppeteer": "^23.10.4",
-    "canvas": "^2.11.2",
     "qrcode": "^1.5.4",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.2",
@@ -131,7 +129,6 @@
     "@types/node": "^20.19.37",
     "@types/passport-jwt": "^4.0.1",
     "@types/passport-local": "^1.0.38",
-    "@types/puppeteer": "^21.0.0",
     "@types/qrcode": "^1.5.6",
     "@types/speakeasy": "^2.0.10",
     "@types/supertest": "^6.0.0",
